Move Data to JSON File
I've read that working with true JSON data is faster/more efficient than working with JavaScript literals. Even if it isn't any faster or any more efficient, it would still be nice to get the 2k+ lines of static data out of my main JavaScript file. This is the bare minimum attempt to get the data out of the main JavaScript file, and into proper JSON format in its own file.