Simplify Drawer Interaction
Instead of having multiple JavaScript toggle()s for multiple CSS classes, I thought it'd be easier to toggle the drawer's visibility and everything else with just one class on the document body.

This commit also fixes two related issues: where the drawer toggle button moved left when the drawer was open, and where the scroll behavior on the document body was not returned to its default state when closing the drawer by clicking on an item.