Add Background Colors to Item Data
Coinciding with the No Man's Sky "Waypoint" (4.0) update, I've added colors to every item currently in my data set. The new design is a vertical 2-color gradient, so I took screenshots of each item in-game, then sampled the colors in Photoshop, and added those hex values here. Given that the way that these colors don't always correlate to the item's type or other predictable datapoint, this seems the best approach for now.