Fixed: Item Data Not Loading on Safari / Firefox
Fixed an issue where the item data was not loading, causing the app to break, on Safari and Firefox. These browsers do not recognize/allow the new "import" JavaScript function, at least not with optional parameters, which are necessary in this case to get the import to accept a JSON file. So, as a workaround, I'm importing the contents of that item data JSON file into `index.php` as inline script.

Fixes #3.