Add Minification to CSS and JS
Adds the gulp-clean-css and gulp-uglify plugins, for minifying CSS and JavaScript files, respectively. Also adds the tasks to the Gulp file to invoke these plugins, updates the index file to load these minified assets instead of the originals, includes the initial versions of these optimized assets, and updates the README to reflect these changes.