Add gulp-tinypng-compress Plugin
Adds the gulp-tinypng-compress plugin for Gulp, which automates the process of compressing PNG images using TinyPNG.com, via their developer API. Also adds a corresponding Gulp task, and adds that task to the default watch process.

I'll be looking at ways to obfuscate this API key next, so I'm going to push it for now, and then request a new key.