Add dotenv
Adds dotenv, a Node Module for managing credentials and other private pieces of information. I'm using it to pass my TinyPNG API key from the `.env` file in the project root (which has been added to the .gitignore file), into the Gulp file where the API key is needed for the gulp-tinypng-compress plugin to work properly.