Implement Vue Router
MVP for this major goal has been met -- Routing and browser history are working, but I still need to add content for a couple new views.

I rewrote some interactive components of the app to make use of Vue Router, which adds browser history for each view change, and makes "going back" to previous views possible. This also had the unintended side effect of simplifying the via lot of my logic. I was able to get rid of a couple components in favor of making them views, as well as removing the process of emitting and passing props between some components in order to maintain state.

This commit closes issues #6 and #7 .