Make Icons into SFCs
I'm following Vue's lead on this one, and copying the SVG code for my icons into individual Vue SFCs. I looked into making this a little less complicated (eg, just pulling in the contents of the SVG files like I've done before with PHP), but that's gonna be a battle for another day.