Filter Translations By Language
Added a set of buttons with one button for each language, and the logic needed to return only those translations which match the text in the input AND the language that corresponds to the selected button.

I also combined the four language data JSON files into a single array in the component's data, so I could make use of the `v-for` directive and cut down on redundant code.