Improvements to Translator View
A lot of translations don't have a third (all caps) word, and a few translations are missing one or two other forms, so I've put each form behind its own `v-if` directive to hide it from view on mobile devices. Desktop users will continue to see an empty space in the table layout.

I've also added a call to `.toUpperCase()` for the all caps words, so they display as they would in-game.